+++ CHANGELOG.md
@@ -32,3 +32,6 @@ Initial stable release
 - #105: Adding placeholder support.
 - #59: Mentions should get deleted on whole by single backspace.
 - #61: Unlink should not get highlighted for mentions.
+
+## 7/12/2016 (1.2.2)
+- #56: Improvements in mention dropdown.